% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{anchorIdx}
\alias{anchorIdx}
\alias{anchorIdx,woven-method}
\title{Indices of the anchor (fully-observed) subjects}
\usage{
anchorIdx(x)

\S4method{anchorIdx}{woven}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{woven}} object}
}
\value{
integer vector
}
\description{
Indices of the anchor (fully-observed) subjects
}
\examples{
data(woven_example)
fit <- woven(woven_example$X_complete, Y = woven_example$Y, K = 3L)
length(anchorIdx(fit))
}
