% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_woven_inputs}
\alias{check_woven_inputs}
\title{Validate inputs before solve}
\usage{
check_woven_inputs(X_list, anchor_idx, K)
}
\arguments{
\item{X_list}{list of modality matrices (n_v x p_v each)}

\item{anchor_idx}{integer vector of anchor indices}

\item{K}{number of latent dimensions}
}
\description{
Validate inputs before solve
}
\keyword{internal}
