% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woven.R
\name{show,woven-method}
\alias{show,woven-method}
\title{Show method for WOVEN fit}
\usage{
\S4method{show}{woven}(object)
}
\arguments{
\item{object}{a woven object from [woven()]}
}
\value{
Invisibly returns \code{NULL}; called for its side effect.
}
\description{
Called automatically when a \code{\linkS4class{woven}} object is printed
or auto-printed at the console.
}
\examples{
set.seed(1)
n <- 20
K <- 2L
X1 <- matrix(rnorm(n * 5), n, 5)
X2 <- matrix(rnorm(n * 4), n, 4)
Y <- rep(1:2, each = n / 2)
miss <- matrix(FALSE, n, 2)
miss[c(15, 17, 19), 1] <- TRUE
miss[c(16, 18, 20), 2] <- TRUE
X1[miss[, 1], ] <- NA
X2[miss[, 2], ] <- NA
anchor_idx <- which(rowSums(miss) == 0)
fit <- woven(list(X1, X2), Y = Y, anchor_idx = anchor_idx, K = K)
fit
}
