% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{singularValues}
\alias{singularValues}
\alias{singularValues,woven-method}
\title{Supervised canonical correlations}
\usage{
singularValues(x)

\S4method{singularValues}{woven}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{woven}} object}
}
\value{
numeric vector of length K
}
\description{
Supervised canonical correlations
}
\examples{
data(woven_example)
fit <- woven(woven_example$X_complete, Y = woven_example$Y, K = 3L)
singularValues(fit)
}
