% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{WList}
\alias{WList}
\alias{WList,woven-method}
\title{Per-modality projection matrices}
\usage{
WList(x)

\S4method{WList}{woven}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{woven}} object}
}
\value{
named list of V numeric matrices, each p_v x K
}
\description{
Per-modality projection matrices
}
\examples{
data(woven_example)
fit <- woven(woven_example$X_complete, Y = woven_example$Y, K = 3L)
names(WList(fit))
}
