% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{woven_ess_retention}
\alias{woven_ess_retention}
\title{Effective sample size retention}
\usage{
woven_ess_retention(n_used, n_total)
}
\arguments{
\item{n_used}{integer, number of subjects with a latent score}

\item{n_total}{integer, total subjects in dataset}
}
\value{
scalar in [0, 1], higher is better (DIABLO structurally caps at overlap fraction)
}
\description{
Effective sample size retention
}
\examples{
woven_ess_retention(n_used = 80, n_total = 100)
}
