% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{woven_nmi}
\alias{woven_nmi}
\title{Normalized mutual information between cluster assignments and true labels}
\usage{
woven_nmi(Z, labels, n_cl = NULL, n_start = 10L)
}
\arguments{
\item{Z}{numeric matrix n x K}

\item{labels}{integer or factor of length n (true labels)}

\item{n_cl}{integer, number of clusters (default = number of unique labels)}

\item{n_start}{integer, k-means random starts}
}
\value{
scalar in [0, 1], higher is better
}
\description{
Uses k-means on Z to get cluster assignments, then computes NMI.
k-means run 10 times to reduce initialization variance.
}
\examples{
set.seed(1)
Z <- matrix(rnorm(40 * 2), 40, 2)
labels <- rep(1:2, each = 20)
woven_nmi(Z, labels)
}
