% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_v2.R
\name{woven_vip}
\alias{woven_vip}
\title{VIP scores for a fitted WOVEN V=2 model}
\usage{
woven_vip(fit)
}
\arguments{
\item{fit}{output of woven_v2}
}
\value{
list with $vip1 (p1-vector) and $vip2 (p2-vector)
}
\description{
VIP scores for a fitted WOVEN V=2 model
}
