% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{Z}
\alias{Z}
\alias{Z,woven-method}
\title{Consensus latent scores}
\usage{
Z(x)

\S4method{Z}{woven}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{woven}} object}
}
\value{
numeric matrix, n x K
}
\description{
Consensus latent scores
}
\examples{
data(woven_example)
fit <- woven(woven_example$X_complete, Y = woven_example$Y, K = 3L)
dim(Z(fit))
}
