% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woven.R
\name{print.woven}
\alias{print.woven}
\title{Print method for WOVEN fit}
\usage{
\method{print}{woven}(x, ...)
}
\arguments{
\item{x}{a woven object from [woven()]}

\item{...}{further arguments (unused)}
}
\description{
Print method for WOVEN fit
}
