% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_als.R
\name{woven_als_vip}
\alias{woven_als_vip}
\title{VIP scores for a fitted ALS model}
\usage{
woven_als_vip(fit)
}
\arguments{
\item{fit}{output of woven_als}
}
\value{
list of V VIP vectors, one per modality
}
\description{
VIP scores for a fitted ALS model
}
