% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_v2.R
\name{woven_vip}
\alias{woven_vip}
\title{VIP scores for a fitted WOVEN V=2 model}
\usage{
woven_vip(fit)
}
\arguments{
\item{fit}{output of woven_v2}
}
\value{
list with $vip1 (p1-vector) and $vip2 (p2-vector)
}
\description{
VIP scores for a fitted WOVEN V=2 model
}
\examples{
set.seed(1)
n <- 20
K <- 2L
X1 <- matrix(rnorm(n * 5), n, 5)
X2 <- matrix(rnorm(n * 4), n, 4)
Y <- rep(1:2, each = n / 2)
miss <- matrix(FALSE, n, 2)
miss[c(15, 17, 19), 1] <- TRUE
miss[c(16, 18, 20), 2] <- TRUE
X1[miss[, 1], ] <- NA
X2[miss[, 2], ] <- NA
anchor_idx <- which(rowSums(miss) == 0)
fit <- woven_v2(X1, X2, anchor_idx = anchor_idx, Y = Y, K = K)
woven_vip(fit)
}
